+++ src/Console/ListScheduler.php
@@ -2,9 +2,13 @@
 
 namespace Hmazter\LaravelScheduleList\Console;
 
+use Carbon\Carbon;
+use Cron\CronExpression;
 use Illuminate\Console\Command;
+use Illuminate\Console\Parser;
 use Illuminate\Console\Scheduling\Event;
 use Illuminate\Console\Scheduling\Schedule;
+use Symfony\Component\Console\Exception\CommandNotFoundException;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
@@ -102,28 +106,73 @@ class ListScheduler extends Command
         $rows = [];
         /** @var Event $event */
         foreach ($events as $event) {
-            $command = $event->buildCommand();
-            $desc = $event->getSummaryForDisplay();
-
-            // if php binary is present in string, it is a the actual command and not a description
-            if (strpos($desc, PHP_BINARY) !== false) {
-                $desc = '';
-            }
-
-            // remove php binary and std output from the command string
-            if ($this->output->getVerbosity() === OutputInterface::VERBOSITY_NORMAL) {
-                $command = substr($command, 0, strpos($command, '>'));
-                $command = trim(str_replace([PHP_BINARY, 'artisan', '\'', '"'], '', $command));
+            $fullCommand = $event->buildCommand();
+            $shortCommand = $this->getShortCommand($fullCommand);
+            $isVerbosityNormal = $this->output->getVerbosity() === OutputInterface::VERBOSITY_NORMAL;
+            list($commandName) = Parser::parse($shortCommand);
+            $description = $event->description;
+            if (empty($description)) {
+                $description = $this->getDescriptionFromCommand($commandName);
             }
 
             $rows[] = [
                 'expression' => $event->getExpression(),
-                'command' => $command,
-                'description' => $desc,
+                'next run at' => $this->getNextRunDate($event),
+                'command' => $isVerbosityNormal ? $shortCommand : $fullCommand,
+                'description' => $description,
             ];
         }
 
         $headers = array_keys($rows[0]);
         $this->table($headers, $rows);
+    }
+
+    /**
+     * Get the "short" command
+     * Remove php binary, "artisan" and std output from the command string
+     *
+     * @param string $command
+     * @return string
+     */
+    private function getShortCommand($command)
+    {
+        $command = substr($command, 0, strpos($command, '>'));
+        $command = trim(str_replace([PHP_BINARY, 'artisan', '\'', '"'], '', $command));
+        return $command;
+    }
+
+    /**
+     * Parse the description from the Command instead of the scheduled event
+     *
+     * @param string $commandName
+     * @return string
+     */
+    private function getDescriptionFromCommand($commandName)
+    {
+        try {
+            $className = get_class($this->getApplication()->find($commandName));
+            $reflection = new \ReflectionClass($className);
+            return (string)$reflection->getDefaultProperties()['description'];
+        } catch (CommandNotFoundException $e) {
+            return '';
+        }
+    }
+
+    /**
+     * Get the next scheduled run date for this event
+     *
+     * @param Event $event
+     * @return string
+     */
+    private function getNextRunDate($event)
+    {
+        $cron = CronExpression::factory($event->getExpression());
+        $date = Carbon::now();
+
+        if ($event->timezone) {
+            $date->setTimezone($event->timezone);
+        }
+
+        return $cron->getNextRunDate()->format('Y-m-d H:i:s');
     }
 }