+++ tests/Integration/TestCommand.php
@@ -1,0 +1,31 @@
+<?php
+
+use Illuminate\Console\Command;
+
+class TestCommand extends Command
+{
+    /**
+     * The name and signature of the console command.
+     *
+     * @var string
+     */
+    protected $signature = 'test:command:two';
+
+    /**
+     * The console command description.
+     *
+     * @var string
+     */
+    protected $description = 'Description of test command';
+
+
+    /**
+     * Execute the console command.
+     *
+     * @return mixed
+     */
+    public function handle()
+    {
+        //
+    }
+}