+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project tries to adhere to [Semantic Versioning](http://semver.org/).
 
+## [0.1.6] - 
+### Updated
+- Output table include next run occasion for each command
+- Description is now parsed from the command class, if available
+
 ## [0.1.5] - 2017-08-12
 ### Added
 - Support Laravel auto discover